Nuevas variables a Usuario añadidas. funciones Scan Implementadas. Solucion de Inicio de main.

+++ src/main/java/DAO/UsuarioDAO.java
@@ -4,6 +4,6 @@ import main.java.usuarios.Usuario;
 
 public interface UsuarioDAO extends GenericDAO<Usuario> {
 
-	public abstract Usuario findByDNI(Long dni);
+	public abstract Usuario findByDNI(int dni);
 	
 }
